requirements.txt
asgiref==3.7.2
autopep8==2.0.2
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.4
django-appconf==1.0.5
django-auditlog==2.3.0
django-cors-headers==4.3.1
django-ranged-response==0.2.0
django-select2==8.1.2
django-serverside-datatable==2.1.0
django-simple-captcha==0.5.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
et-xmlfile==1.1.0
idna==3.4
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mysqlclient==2.2.0
openpyxl==3.1.2
packaging==23.1
Pillow==10.0.0
pycodestyle==2.11.0
PyJWT==2.8.0
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.4
